requirements.txt
flask>=2.0
flask-cors>=3.0
gunicorn>=20.0
psycopg2-binary>=2.9
